# specs/001-scholarship-mvp/contracts/openapi.yaml
openapi: 3.0.3
info:
  title: Scholarship AI Assistant API
  version: "0.1.0"
  description: >
    API contracts for the MVP feature (specs/001-scholarship-mvp). Endpoints are
    grouped by capability per blueprint §5.1; each group is tagged with the
    blueprint phase that implements it. Auth endpoints (/auth/*) are Phase 0,
    already implemented — included here for completeness only, unchanged.
servers:
  - url: /
...
paths:
...
  /auth/signup:
    post:
      tags: [auth]
      summary: Create an account
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/SignupRequest' }
      responses:
        '201':
          description: Account created
          content:
            application/json:
              schema: { $ref: '#/components/schemas/UserResponse' }
        '409': { description: Email already registered }
  /auth/login:
    post:
      tags: [auth]
      summary: Log in
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/LoginRequest' }
      responses:
        '200':
          description: Token issued
          content:
            application/json:
              schema: { $ref: '#/components/schemas/TokenResponse' }
        '401': { description: Incorrect email or password }
...
  /profile:
    get:
      tags: [profile]
      summary: Get the current user's living profile
      security: [{ bearerAuth: [] }]
      responses:
        '200':
          description: Profile (including criteria classification and missing-info list)
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Profile' }
        '404': { description: No profile yet created for this user }
    put:
      tags: [profile]
      summary: Create or update the profile (partial updates supported — living profile, not a one-time form)
      security: [{ bearerAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/ProfileUpdateRequest' }
      responses:
        '200':
          description: Updated profile
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Profile' }
  /profile/criteria:
    post:
      tags: [profile]
      summary: Add or reclassify a profile criterion (hard_constraint / soft_preference / exclusion)
      security: [{ bearerAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/ProfileCriterionRequest' }
      responses:
        '200':
          description: Updated criterion list
          content:
            application/json:
              schema:
                type: array
                items: { $ref: '#/components/schemas/ProfileCriterion' }
        '422':
          description: Invalid criterion kind
  /profile/import-cv:
    post:
      tags: [profile]
      summary: Extract profile fields from an uploaded CV (extract_profile tool; suggestions only, user confirms before save)
      security: [{ bearerAuth: [] }]
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required: [file]
              properties:
                file: { type: string, format: binary }
      responses:
        '200':
          description: Extracted field suggestions (not yet persisted)
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ProfileUpdateRequest' }

  # ---------------------------------------------------------------------
  # Phase 1 — Scholarship URL Matching (FR-URL-1/2)
  # ---------------------------------------------------------------------
...
  /discovery:
    post:
      tags: [discovery]
      summary: Find candidate scholarships from governed sources matching the profile
      security: [{ bearerAuth: [] }]
      responses:
        '200':
          description: Ranked, source-attributed results plus a measured coverage summary
          content:
            application/json:
              schema: { $ref: '#/components/schemas/DiscoveryResult' }
        '422':
          description: Profile setup is required before discovery can run

  # ---------------------------------------------------------------------
  # Phase 3 — Documents (FR-DOC-1..3)
  # ---------------------------------------------------------------------
...
  /applications/{application_id}/generate/cv:
    post:
      tags: [generation]
      summary: Generate a requirement-aware CV grounded strictly in profile + uploaded documents
      security: [{ bearerAuth: [] }]
      parameters:
        - name: application_id
          in: path
          required: true
          schema: { type: string, format: uuid }
      responses:
        '200':
          description: Generated CV with full source trace
          content:
            application/json:
              schema: { $ref: '#/components/schemas/GeneratedDocument' }
        '409':
          description: Gap in required information — reported, not filled with invented content (US5 Scenario 5)
  /applications/{application_id}/generate/sop:
    post:
      tags: [generation]
      summary: Generate a requirement-aware SOP/motivation essay answering the university's specified questions
      security: [{ bearerAuth: [] }]
      parameters:
        - name: application_id
          in: path
          required: true
          schema: { type: string, format: uuid }
      responses:
        '200':
          description: Generated SOP with full source trace
          content:
            application/json:
              schema: { $ref: '#/components/schemas/GeneratedDocument' }
        '409':
          description: Gap in required information — reported, not filled with invented content

  # ---------------------------------------------------------------------
  # Phase 4 — Application Planner / Tracker / Assistant / Submission Approval
  # ---------------------------------------------------------------------
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    SignupRequest:
      type: object
      required: [email, password]
      properties:
        email: { type: string, format: email }
        password: { type: string }
    LoginRequest:
      type: object
      required: [email, password]
      properties:
        email: { type: string, format: email }
        password: { type: string }
    TokenResponse:
      type: object
      properties:
        access_token: { type: string }
        token_type: { type: string, default: bearer }
    UserResponse:
      type: object
      properties:
        id: { type: string, format: uuid }
        email: { type: string, format: email }
        created_at: { type: string, format: date-time }

    # ---- Shared value-status vocabulary (constitution Principle I) ----
...
    ProfileCriterion:
      type: object
      properties:
        id: { type: string, format: uuid }
        dimension: { type: string }
        operator: { type: string }
        value: {}
        kind: { type: string, enum: [hard_constraint, soft_preference, exclusion] }
        weight: { type: number, nullable: true }
        note: { type: string, nullable: true }
    ProfileCriterionRequest:
      allOf:
        - $ref: '#/components/schemas/ProfileCriterion'
      required: [dimension, operator, kind]
    MissingInfoItem:
      type: object
      properties:
        dimension: { type: string }
        reason: { type: string }
    Profile:
      type: object
      properties:
        id: { type: string, format: uuid }
        name: { type: string, nullable: true }
        nationality: { type: string, nullable: true }
        country_of_residence: { type: string, nullable: true }
        current_degree: { type: string, nullable: true }
        target_degree_level: { type: string, enum: [BS, MS, PhD, other], nullable: true }
        target_fields: { type: array, items: { type: string } }
        target_countries: { type: array, items: { type: string } }
        education_records: { type: array, items: { $ref: '#/components/schemas/EducationRecordOut' } }
        test_scores: { type: array, items: { $ref: '#/components/schemas/TestScoreOut' } }
        experience: { type: array, items: { $ref: '#/components/schemas/ExperienceOut' } }
        criteria: { type: array, items: { $ref: '#/components/schemas/ProfileCriterion' } }
        missing_info: { type: array, items: { $ref: '#/components/schemas/MissingInfoItem' } }
        updated_at: { type: string, format: date-time }
    ProfileUpdateRequest:
      type: object
      description: >
        Partial update — any subset of Profile's writable fields (living profile, not a
        one-time form). `education_records`/`test_scores`/`experience`, when present,
        replace that satellite table's rows wholesale for this profile.
      properties:
        name: { type: string }
        nationality: { type: string }
        country_of_residence: { type: string }
        current_degree: { type: string }
        target_degree_level: { type: string, enum: [BS, MS, PhD, other], nullable: true }
        target_fields: { type: array, items: { type: string } }
        target_countries: { type: array, items: { type: string } }
        education_records: { type: array, items: { $ref: '#/components/schemas/EducationRecordIn' } }
        test_scores: { type: array, items: { $ref: '#/components/schemas/TestScoreIn' } }
        experience: { type: array, items: { $ref: '#/components/schemas/ExperienceIn' } }
    EducationRecordIn:
      type: object
      properties:
        degree: { type: string, nullable: true }
        field: { type: string, nullable: true }
        university: { type: string, nullable: true }
        gpa: { type: number, nullable: true }
        gpa_scale: { type: number, nullable: true }
        start: { type: string, format: date, nullable: true }
        end: { type: string, format: date, nullable: true }
    EducationRecordOut:
      allOf:
        - $ref: '#/components/schemas/EducationRecordIn'
        - type: object
          required: [id]
          properties:
            id: { type: string, format: uuid }
    TestType:
      type: string
      enum: [IELTS, TOEFL, PTE, GRE, GMAT, other]
    TestStatus:
      type: string
      enum: [have, planned, none]
    TestScoreIn:
      type: object
      required: [test_type, status]
      properties:
        test_type: { $ref: '#/components/schemas/TestType' }
        status: { $ref: '#/components/schemas/TestStatus' }
        score: { type: number, nullable: true }
        taken_at: { type: string, format: date, nullable: true }
    TestScoreOut:
      allOf:
        - $ref: '#/components/schemas/TestScoreIn'
        - type: object
          required: [id]
          properties:
            id: { type: string, format: uuid }
    ExperienceKind:
      type: string
      enum: [work, research, publication, project]
    ExperienceIn:
      type: object
      required: [kind]
      properties:
        kind: { $ref: '#/components/schemas/ExperienceKind' }
        title: { type: string, nullable: true }
        org: { type: string, nullable: true }
        detail: { type: string, nullable: true }
        start: { type: string, format: date, nullable: true }
        end: { type: string, format: date, nullable: true }
    ExperienceOut:
      allOf:
        - $ref: '#/components/schemas/ExperienceIn'
        - type: object
          required: [id]
          properties:
            id: { type: string, format: uuid }

    # ---- Shared: scholarship field / evidence ----
...
    CoverageSummary:
      type: object
      properties:
        sources_checked: { type: integer }
        sources_failed: { type: integer }
        countries_covered: { type: array, items: { type: string } }
        gaps: { type: array, items: { type: string } }
        claims_complete_coverage: { type: boolean, enum: [false], description: "MUST always be false — FR-DISC-7" }
    DiscoveryResultItem:
      type: object
      properties:
        scholarship_id: { type: string, format: uuid }
        source_id: { type: string, format: uuid }
        source_type: { type: string }
        verification_status: { type: string, enum: [verified, unverified, conflicting] }
    DiscoveryResult:
      type: object
      properties:
        results: { type: array, items: { $ref: '#/components/schemas/DiscoveryResultItem' } }
        coverage: { $ref: '#/components/schemas/CoverageSummary' }

    # ---- Phase 3: Documents / Generation ----
...
    GeneratedDocument:
      type: object
      properties:
        id: { type: string, format: uuid }
        type: { type: string, enum: [cv, sop, motivation, other] }
        file_ref: { type: string }
        source_trace:
          type: array
          items:
            type: object
            properties:
              claim: { type: string }
              grounded_in: { type: string, description: "profile field or document id" }

    # ---- Phase 4: Application / Planner / Tracker / Assistant / Submission ----
...